ValidationReport as a separate class
